// tools/motion/motion_test_watcher_app/src/model/constants.ts
export enum AnimationEventPhasename {
    DONE = 'done'
}

export enum AnimationEventTime {
    ENTRY_TIME = 300
}

export enum TestListPanelDimensionsFactor {
    DEFAULT_WIDTH = 0.25,
    MIN_WIDTH = 0.15,
    MAX_WIDTH = 0.6
}

export enum VideoPanelDimensionsFactor {
    DEFAULT_HEIGHT = 0.5,
    MIN_HEIGHT = 0.1,
    MAX_HEIGHT = 0.9
}

export const DIVIDER_HEIGHT = 8; // 8px for the h-2 divider height
export const MIN_PANEL_WIDTH = 100;
export const MIN_PANEL_HEIGHT = 100;